+++ crates/rust-client/src/lib.rs
@@ -196,8 +196,7 @@ pub mod asset {
 pub mod auth {
     pub use miden_lib::AuthScheme;
     pub use miden_lib::account::auth::{AuthRpoFalcon512, NoAuth};
-    pub use miden_objects::account::auth::PublicKeyCommitment;
-    pub use miden_objects::account::{AuthSecretKey, Signature};
+    pub use miden_objects::account::auth::{AuthSecretKey, PublicKeyCommitment, Signature};
     pub use miden_tx::auth::{BasicAuthenticator, SigningInputs, TransactionAuthenticator};
 }
 
@@ -236,7 +235,13 @@ pub mod crypto {
 
 /// Provides types for working with addresses within the Miden network.
 pub mod address {
-    pub use miden_objects::address::{AccountIdAddress, Address, AddressInterface, NetworkId};
+    pub use miden_objects::address::{
+        Address,
+        AddressId,
+        AddressInterface,
+        NetworkId,
+        RoutingParameters,
+    };
 }
 
 /// Provides types for working with the virtual machine within the Miden network.