+++ crates/web-client/test/notes.test.ts
@@ -253,7 +253,9 @@ test.describe("createP2IDNote and createP2IDENote", () => {
         let outputNote = window.OutputNote.full(p2IdNote);
 
         let transactionRequest = new window.TransactionRequestBuilder()
-          .withOwnOutputNotes(new window.OutputNotesArray([outputNote]))
+          .withOwnOutputNotes(
+            new window.MidenArrays.OutputNoteArray([outputNote])
+          )
           .build();
 
         let transactionUpdate = await window.helpers.executeAndApplyTransaction(
@@ -354,7 +356,9 @@ test.describe("createP2IDNote and createP2IDENote", () => {
         let outputNote = window.OutputNote.full(p2IdeNote);
 
         let transactionRequest = new window.TransactionRequestBuilder()
-          .withOwnOutputNotes(new window.OutputNotesArray([outputNote]))
+          .withOwnOutputNotes(
+            new window.MidenArrays.OutputNoteArray([outputNote])
+          )
           .build();
 
         let transactionUpdate = await window.helpers.executeAndApplyTransaction(